+++ app/src/main/java/com/ostah_app/data/remote/objects/OrderTecket.kt
@@ -32,9 +32,9 @@ data class OrderTecket (
     @SerializedName("ticket_type_id")
     var ticket_type_id:Int,
     @SerializedName("lat")
-    var lat:Double,
+    var lat:Float,
     @SerializedName("lng")
-    var lng:Double,
+    var lng:Float,
     @SerializedName("reformer")
     var reformer:UserResposeModel,
     @SerializedName("status")
